OstiaMaredeDeuPush

+++ DWES/Tema4/Projecte_Tema4/borrarmapa.php
@@ -1,4 +1,7 @@
 <?php
+
+use GuzzleHttp\BodySummarizer;
+
 include_once __DIR__ . '/models/MapaDAO.php';
 include_once __DIR__ . '/models/Mapa.php';
 include_once __DIR__. '/models/Territori.php';
@@ -9,21 +12,28 @@ include_once __DIR__ . '/functions.php';
 session_start();
 //Si no está loguejat el redirigim a login
 if (!isset($_SESSION['usuari'])) {
-    header('Location: index.php');
+        header('Location: index.php');
     exit;
   }
 $id="";
 
-if($_SERVER['REQUEST_METHOD']=="GET"){
-    if(isset($_GET["id"])){
-        $id=neteja_dades($_GET["id"]);
-        $mapa=MapaDao::select($id);
+if($_SERVER['REQUEST_METHOD']=="GET"&& isset($_GET["id"])){
+
+    $id=neteja_dades($_GET["id"]);
+    $mapa=MapaDao::select($id);
+
+    if ($mapa) {
         TerritoriDao::deleteTerritorisByMapaId($id);
-        if($mapa &&MapaDao::delete($mapa)){
-            $_SESSION['missatge_borrat']='Mapa eliminat correctament';
-            header("location:index.php");
-            exit;
+
+        if(MapaDao::delete($mapa)){
+            $_SESSION['missatge_borrat']='Mapa i territoris associats eliminats correctament';
+            echo("Mapa borrat");
+            //header('Location: index.php');
+            //exit;
         }
-    }
+    }  
 }
+// Si no hem pogut eliminar o hi ha hagut un error, redirigim al index
+header('Location: index.php');
+exit;
 ?>
